// frontend/src/apis/pipeline/api.ts
export interface V1Parameter {
  /**
   *
   * @type {string}
   * @memberof V1Parameter
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof V1Parameter
   */
  value?: string;
}
// ...
export interface V1PipelineVersion {
  /**
   * Output. Unique version ID. Generated by API server.
   * @type {string}
   * @memberof V1PipelineVersion
   */
  id?: string;
  /**
   * Optional input field. Version name provided by user.
   * @type {string}
   * @memberof V1PipelineVersion
   */
  name?: string;
  /**
   * Output. The time this pipeline version is created.
   * @type {Date}
   * @memberof V1PipelineVersion
   */
  created_at?: Date;
  /**
   * Output. The input parameters for this pipeline.
   * @type {Array<V1Parameter>}
   * @memberof V1PipelineVersion
   */
  parameters?: Array<V1Parameter>;
  /**
   * Input. Optional. Pipeline version code source.
   * @type {string}
   * @memberof V1PipelineVersion
   */
  code_source_url?: string;
  /**
   * Input. Required. Pipeline version package url. Whe calling CreatePipelineVersion API method, need to provide one package file location.
   * @type {V1Url}
   * @memberof V1PipelineVersion
   */
  package_url?: V1Url;
  /**
   * Input field. Specify which resource this pipeline version belongs to. For Experiment, the only valid resource reference is a single Namespace.
   * @type {Array<V1ResourceReference>}
   * @memberof V1PipelineVersion
   */
  resource_references?: Array<V1ResourceReference>;
  /**
   * Input. Optional. Description for the pipeline version.
   * @type {string}
   * @memberof V1PipelineVersion
   */
  description?: string;
}

/**
 *
 * @export
 * @enum {string}
 */
export enum V1Relationship {
  UNKNOWNRELATIONSHIP = <any>'UNKNOWN_RELATIONSHIP',
  OWNER = <any>'OWNER',
  CREATOR = <any>'CREATOR',
}

/**
 *
 * @export
 * @interface V1ResourceKey
 */
export interface V1ResourceKey {
  /**
   * The type of the resource that referred to.
   * @type {V1ResourceType}
   * @memberof V1ResourceKey
   */
  type?: V1ResourceType;
  /**
   * The ID of the resource that referred to.
   * @type {string}
   * @memberof V1ResourceKey
   */
  id?: string;
}

/**
 *
 * @export
 * @interface V1ResourceReference
 */
export interface V1ResourceReference {
  /**
   *
   * @type {V1ResourceKey}
   * @memberof V1ResourceReference
   */
  key?: V1ResourceKey;
  /**
   * The name of the resource that referred to.
   * @type {string}
   * @memberof V1ResourceReference
   */
  name?: string;
  /**
   * Required field. The relationship from referred resource to the object.
   * @type {V1Relationship}
   * @memberof V1ResourceReference
   */
  relationship?: V1Relationship;
}
// ...
export interface V1Url {
  /**
   * URL of the pipeline definition or the pipeline version definition.
   * @type {string}
   * @memberof V1Url
   */
  pipeline_url?: string;
}