// frontend/src/apis/run/api.ts
export interface PipelineSpecRuntimeConfig {
  /**
   * The runtime parameters of the PipelineSpec. The parameters will be used to replace the placeholders at runtime.
   * @type {{ [key: string]: V1Value; }}
   * @memberof PipelineSpecRuntimeConfig
   */
  parameters?: { [key: string]: V1Value };
  /**
   *
   * @type {string}
   * @memberof PipelineSpecRuntimeConfig
   */
  pipeline_root?: string;
}
// ...
export enum RunMetricFormat {
  UNSPECIFIED = <any>'UNSPECIFIED',
  RAW = <any>'RAW',
  PERCENTAGE = <any>'PERCENTAGE',
}
// ...
export interface V1Parameter {
  /**
   *
   * @type {string}
   * @memberof V1Parameter
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof V1Parameter
   */
  value?: string;
}

/**
 *
 * @export
 * @interface V1PipelineRuntime
 */
export interface V1PipelineRuntime {
  /**
   * Output. The runtime JSON manifest of the pipeline, including the status of pipeline steps and fields need for UI visualization etc.
   * @type {string}
   * @memberof V1PipelineRuntime
   */
  pipeline_manifest?: string;
  /**
   * Output. The runtime JSON manifest of the argo workflow. This is deprecated after pipeline_runtime_manifest is in use.
   * @type {string}
   * @memberof V1PipelineRuntime
   */
  workflow_manifest?: string;
}

/**
 *
 * @export
 * @interface V1PipelineSpec
 */
export interface V1PipelineSpec {
  /**
   * Optional input field. The ID of the pipeline user uploaded before.
   * @type {string}
   * @memberof V1PipelineSpec
   */
  pipeline_id?: string;
  /**
   * Optional output field. The name of the pipeline. Not empty if the pipeline id is not empty.
   * @type {string}
   * @memberof V1PipelineSpec
   */
  pipeline_name?: string;
  /**
   * Optional input field. The marshalled raw argo JSON workflow. This will be deprecated when pipeline_manifest is in use.
   * @type {string}
   * @memberof V1PipelineSpec
   */
  workflow_manifest?: string;
  /**
   * Optional input field. The raw pipeline JSON spec.
   * @type {string}
   * @memberof V1PipelineSpec
   */
  pipeline_manifest?: string;
  /**
   *
   * @type {Array<V1Parameter>}
   * @memberof V1PipelineSpec
   */
  parameters?: Array<V1Parameter>;
  /**
   *
   * @type {PipelineSpecRuntimeConfig}
   * @memberof V1PipelineSpec
   */
  runtime_config?: PipelineSpecRuntimeConfig;
}

/**
 *
 * @export
 * @interface V1ReadArtifactResponse
 */
export interface V1ReadArtifactResponse {
  /**
   * The bytes of the artifact content.
   * @type {string}
   * @memberof V1ReadArtifactResponse
   */
  data?: string;
}

/**
 *
 * @export
 * @enum {string}
 */
export enum V1Relationship {
  UNKNOWNRELATIONSHIP = <any>'UNKNOWN_RELATIONSHIP',
  OWNER = <any>'OWNER',
  CREATOR = <any>'CREATOR',
}
// ...
export interface V1ResourceKey {
  /**
   * The type of the resource that referred to.
   * @type {V1ResourceType}
   * @memberof V1ResourceKey
   */
  type?: V1ResourceType;
  /**
   * The ID of the resource that referred to.
   * @type {string}
   * @memberof V1ResourceKey
   */
  id?: string;
}

/**
 *
 * @export
 * @interface V1ResourceReference
 */
export interface V1ResourceReference {
  /**
   *
   * @type {V1ResourceKey}
   * @memberof V1ResourceReference
   */
  key?: V1ResourceKey;
  /**
   * The name of the resource that referred to.
   * @type {string}
   * @memberof V1ResourceReference
   */
  name?: string;
  /**
   * Required field. The relationship from referred resource to the object.
   * @type {V1Relationship}
   * @memberof V1ResourceReference
   */
  relationship?: V1Relationship;
}

/**
 *
 * @export
 * @enum {string}
 */
export enum V1ResourceType {
  UNKNOWNRESOURCETYPE = <any>'UNKNOWN_RESOURCE_TYPE',
  EXPERIMENT = <any>'EXPERIMENT',
  JOB = <any>'JOB',
  PIPELINE = <any>'PIPELINE',
  PIPELINEVERSION = <any>'PIPELINE_VERSION',
  NAMESPACE = <any>'NAMESPACE',
}

/**
 *
 * @export
 * @interface V1Run
 */
export interface V1Run {
  /**
   * Output. Unique run ID. Generated by API server.
   * @type {string}
   * @memberof V1Run
   */
  id?: string;
  /**
   * Required input field. Name provided by user, or auto generated if run is created by scheduled job. Not unique.
   * @type {string}
   * @memberof V1Run
   */
  name?: string;
  /**
   * Output. Specify whether this run is in archived or available mode.
   * @type {V1RunStorageState}
   * @memberof V1Run
   */
  storage_state?: V1RunStorageState;
  /**
   *
   * @type {string}
   * @memberof V1Run
   */
  description?: string;
  /**
   * Required input field. Describing what the pipeline manifest and parameters to use for the run.
   * @type {V1PipelineSpec}
   * @memberof V1Run
   */
  pipeline_spec?: V1PipelineSpec;
  /**
   * Optional input field. Specify which resource this run belongs to. When creating a run from a particular pipeline version, the pipeline version can be specified here.
   * @type {Array<V1ResourceReference>}
   * @memberof V1Run
   */
  resource_references?: Array<V1ResourceReference>;
  /**
   * Optional input field. Specify which Kubernetes service account this run uses.
   * @type {string}
   * @memberof V1Run
   */
  service_account?: string;
  /**
   * Output. The time that the run created.
   * @type {Date}
   * @memberof V1Run
   */
  created_at?: Date;
  /**
   * Output. When this run is scheduled to run. This could be different from created_at. For example, if a run is from a backfilling job that was supposed to run 2 month ago, the scheduled_at is 2 month ago, v.s. created_at is the current time.
   * @type {Date}
   * @memberof V1Run
   */
  scheduled_at?: Date;
  /**
   * Output. The time this run is finished.
   * @type {Date}
   * @memberof V1Run
   */
  finished_at?: Date;
  /**
   *
   * @type {string}
   * @memberof V1Run
   */
  status?: string;
  /**
   * In case any error happens retrieving a run field, only run ID and the error message is returned. Client has the flexibility of choosing how to handle error. This is especially useful during listing call.
   * @type {string}
   * @memberof V1Run
   */
  error?: string;
  /**
   * Output. The metrics of the run. The metrics are reported by ReportMetrics API.
   * @type {Array<V1RunMetric>}
   * @memberof V1Run
   */
  metrics?: Array<V1RunMetric>;
}

/**
 *
 * @export
 * @interface V1RunDetail
 */
export interface V1RunDetail {
  /**
   *
   * @type {V1Run}
   * @memberof V1RunDetail
   */
  run?: V1Run;
  /**
   *
   * @type {V1PipelineRuntime}
   * @memberof V1RunDetail
   */
  pipeline_runtime?: V1PipelineRuntime;
}

/**
 *
 * @export
 * @interface V1RunMetric
 */
export interface V1RunMetric {
  /**
   * Required. The user defined name of the metric. It must between 1 and 63 characters long and must conform to the following regular expression: `[a-z]([-a-z0-9]*[a-z0-9])?`.
   * @type {string}
   * @memberof V1RunMetric
   */
  name?: string;
  /**
   * Required. The runtime node ID which reports the metric. The node ID can be found in the RunDetail.workflow.Status. Metric with same (node_id, name) are considerd as duplicate. Only the first reporting will be recorded. Max length is 128.
   * @type {string}
   * @memberof V1RunMetric
   */
  node_id?: string;
  /**
   * The number value of the metric.
   * @type {number}
   * @memberof V1RunMetric
   */
  number_value?: number;
  /**
   * The display format of metric.
   * @type {RunMetricFormat}
   * @memberof V1RunMetric
   */
  format?: RunMetricFormat;
}

/**
 *
 * @export
 * @enum {string}
 */
export enum V1RunStorageState {
  AVAILABLE = <any>'STORAGESTATE_AVAILABLE',
  ARCHIVED = <any>'STORAGESTATE_ARCHIVED',
}